% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotme.R
\name{plotme}
\alias{plotme}
\title{A function to generate plots for single shiny interface}
\usage{
plotme(refdata = NULL, sortdata = NULL, power = TRUE,
  absolutevalue = TRUE, ttype = "pm", splitn = NULL, predlevel = 0.9,
  testtype = TRUE)
}
\arguments{
\item{file}{test}

\item{population}{teast}
}
\description{
A function to generate plots for single shiny interface
}
\examples{
plotme()
}
\keyword{random}
